add packaged modules
