+++ marine-examples/records/Config.toml
@@ -1,7 +1,7 @@
 modules_dir = "artifacts/"
 
 [[module]]
-name = "records_effector"
+name = "records_facade"
 max_heap_size = "10  KiB"
 logger_enabled = true
