+++ marine-examples/records/build.sh
@@ -2,17 +2,15 @@
 set -o errexit -o nounset -o pipefail
 
 # This script builds all subprojects and puts all created Wasm modules in one dir
-cd effector
-cargo update --aggressive
+cd pure
 marine build --release
 
-cd ../pure
-cargo update --aggressive
+cd ../facade
 marine build --release
 
 cd ..
 
 mkdir -p artifacts
 rm -f artifacts/*.wasm
-cp effector/target/wasm32-wasi/release/records_effector.wasm artifacts/
 cp pure/target/wasm32-wasi/release/records_pure.wasm artifacts/
+cp facade/target/wasm32-wasi/release/records_facade.wasm artifacts/
