--- marine-examples/records/effector/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "record-effector"
-version = "0.1.0"
-authors = ["Fluence Labs"]
-edition = "2018"
-publish = false
-
-[[bin]]
-name = "records_effector"
-path = "src/main.rs"
-
-[dependencies]
-marine-rs-sdk = "0.7.1"
-test-record = { path = "../test-record" }
